+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `StreamMessageListView.floatingDateDividerBuilder` in order to specifically customize the
   floating date divider.
+- Added spacing to typing indicator.
 
 ## 9.18.0